UCR expand column is generic
